Move Dorfler marking function out of basesolver.cpp and into dedicated file
