+++ palace/utils/CMakeLists.txt
@@ -8,6 +8,7 @@
 target_sources(${TARGET_NAME}
   PRIVATE
   ${CMAKE_CURRENT_SOURCE_DIR}/configfile.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/dorfler.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/errorindicators.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/geodata.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/iodata.cpp
